Adding in NewContextWithStrategy for go binding
Adding in NewContextWithStrategy, a function to allow you to specify the
strategy when initializing the context. NewContext will use this
"helper" function so that users don't need to think about what strategy
they're specifying by default. I'm also doing it this way (instead of a
function that sets the strategy) because the basis of what parameters we
populate the context with is determined by the strategy we are using, so
it needs to be done in the beginning of creating the context.